Fixed Bugs + Updated Analytics

+++ fem-lab/functions/help function/hf_eval_solution_vectorized.m
@@ -77,7 +77,7 @@ function val=hf_eval_solution_vectorized(x,y,u,cell_matrix,vertex_matrix,SF)
         % The diagonal of this matrix consists of the correct pairs of shape function and cell and we can use the inner product with the right entries of u
         % If not for this construction, we would have had to loop over each entry of our active_node vector, evaluate a polynomial and multiply with a scalar.
         % Since matrix/vector operations are optimized in octave, this approach is more efficient.
-        val=u(active_node)'*diag(hf_eval_poly_transformed(x,y,SF(local_node,:),mesh_size,vertex_matrix(cell_matrix(active_cell_vector,1),1),vertex_matrix(cell_matrix(active_cell_vector,1),2)));
+        val=u(int32(active_node))'*diag(hf_eval_poly_transformed(x,y,SF(int32(local_node),:),mesh_size,vertex_matrix(cell_matrix(int32(active_cell_vector),1),1),vertex_matrix(cell_matrix(int32(active_cell_vector),1),2)));
     endif
